Rename test1.csv to test_Y_position_data.csv
This single column of numbers represents the Y position of a fish tracked with idTracker.ai. Originally, the output csv file was composed of two columns: the first column contained the video frame number, and the second contained the Y position of the tracked body centroid. I cut the frame column as well as a few thousand initially untracked frames that displayed a NaN value. The resulting single-column csv is therefore cleaned up and ready to go for analysis using this project's script. Future quality of life updates will roll out as my coding skills improve.

\Note that in computing, the origin of an image is in the top left corner, rather than in the bottom left as is customary in the math we've been taught since elementary school. This means that larger Y values reflect the fish moving down in the tank,